% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-topic.R
\name{get_topic}
\alias{get_topic}
\title{Retrieve Hot Topics (Trends) for a Given Country}
\usage{
get_topic(country = "United States")
}
\arguments{
\item{country_name}{The country name (see \code{\link[=topic_countries]{topic_countries()}})}
}
\value{
data frame
}
\description{
Retrieve Hot Topics (Trends) for a Given Country
}
\examples{
x <- get_topic()
}
