% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-topic.R
\name{topic_countries}
\alias{topic_countries}
\title{Return a character vector of supported countries}
\usage{
topic_countries()
}
\arguments{
\item{pattern}{a valid regular expression used to filter the list}
}
\value{
character vector
}
\description{
Use a regular expression in \code{pattern} to filter the list.
}
\examples{
head(topic_countries())
}
