% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hottopic-package.R
\docType{package}
\name{hottopic}
\alias{hottopic}
\alias{hottopic-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/hottopic}
\item BugReports: \url{https://gitlab.com/hrbrmstr/hottopic/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
